[databroker] Add kuksa.val.v1 protos
