+++ kuksa_databroker/databroker-api/build.rs
@@ -19,6 +19,8 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                 "proto/sdv/databroker/v1/broker.proto",
                 "proto/sdv/databroker/v1/types.proto",
                 "proto/sdv/databroker/v1/collector.proto",
+                "proto/kuksa/val/v1/val.proto",
+                "proto/kuksa/val/v1/types.proto",
             ],
             &["proto"],
         )?;
