+++ kuksa_databroker/proto/kuksa/val/v1/types.proto
@@ -1,0 +1,1 @@
+../../../../../proto/kuksa/val/v1/types.proto
