+++ kuksa_databroker/proto/kuksa/val/v1/val.proto
@@ -1,0 +1,1 @@
+../../../../../proto/kuksa/val/v1/val.proto
